Fix nsss linkage
When enabling nsss, execline fails to compile with the following
message:

	| ./tools/gen-multicall.sh > src/multicall/execline.c
	| make: *** No rule to make target '-lnsss', needed by 'execline'.  Stop.
	| make: *** Waiting for unfinished jobs....

This is due to the fact that 'tools/gen_deps.sh' expects that each
external library variable ends in '_LIB' to add it to the 'EXTRA_LIBS'
variable for the executable target. So, rename the dependency and re-run
'tools/gen-deps.sh' to fix this.

Signed-off-by: Mathieu Anquetin <mathieu.anquetin@groupe-cahors.com>